requirements.txt
numpy==2.1.0
openTSNE==1.0.2
openTSNE==1.0.2
pandas==2.2.2
Pillow==10.4.0
plotly==5.9.0
scikit_learn==1.5.1
streamlit==1.38.0
streamlit_drawable_canvas==0.9.3
